% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project_tree}
\alias{get_project_tree}
\title{Show example project structure as tree}
\usage{
get_project_tree()
}
\value{
data.tree object
}
\description{
Use the data.tree package to show an example
file structure for shiny app package
}
